[FIX] account_asset: allow disposing assets on hashed journals
Currently, trying to sell or dispose an asset on a hashed journal
leads to a `UserError`: "The remaining value on the last depreciation line must be 0".

This occurs because `_reverse_moves()` tries to change the depreciation amount on the
first available draft move. There aren't any since the method is being called right after
removing all the draft depreciation moves. So it creates a new one with a negative
`asset_remaining_value`.

This is fixed by avoiding the creation of said move on assets that are being closed.

A test is added to cover this use case.

closes odoo/enterprise#72819

Related: odoo/odoo#184535
Signed-off-by: William André (wan) <wan@odoo.com>