[FIX] hr_expense_extract: don't crash when searching for an inactive or non-existing currency
When extracting data from a ticket expressed in a currency that's inactive in the company, the code crashed with a KeyError, because the 'currency_id' was not in the vals dictionary.

This crashed some tests in hr_expense_extract when they were run in a db with only USD as active currency.

Part-of: odoo/enterprise#72941
Signed-off-by: William André (wan) <wan@odoo.com>