[FIX] hr_expense_extract: don't crash when searching for an inactive or non-existing currency
When extracting data from a ticket expressed in a currency that's inactive in the company, the code crashed with a KeyError, because the 'currency_id' was not in the vals dictionary.

This crashed some tests in hr_expense_extract when they were run in a db with only USD as active currency.

X-original-commit: 0e4e2b3e36524c97085b9ecd7c29a24c8c8b381f
Part-of: odoo/enterprise#74300
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>