[FIX] l10n_es_reports: fix "Type for mod 347" default value
Steps to reproduce:
- Install l10n_es_reports and Sales
- Create a SO for a Spanish customer
- Create an invoice from the SO
- Check "AEAT data" tab of the invoice

Issue:
"Type for mod 347" is not set.
It should be set to "Regular operation" as it is done when creating
an invoice manually.

Cause:
The default is handled by a function that depends on another field.
However that other field is not set when computing the default
value and the result is always False.

Solution:
There were a function for the default (_default_mod_349_invoice_type)
and an onchange on partner (_onchange_partner_id_set_347_invoice_type).
Both of them were checking different conditions.
The fix is merging these 2 methods by converting the field into a
computed stored field where both conditions are handled.

opw-4194030

closes odoo/enterprise#73508

X-original-commit: fef8a6725451b0a7a63f77e8fe847e4c61e14706
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>