[FIX] web_studio: add data-oe-demo support for t-out
Some reports have been edited to have qweb default content that was
meant to only be used from within studio.

```xml
<t t-field="some_field">studio placeholder</t>
```

This content is displayed if the evaluated value is either False or
None. This can also happen outside studio, if the field is not required.

An attribute `data-oe-demo` was introduced for t-fields.

```xml
<t t-field="some_field" data-oe-demo="studio placeholder"/>
```

This attribute was introduced for t-field nodes in order to only display
placeholder content inside the report editor.

We didn't introduce it for `t-out` as it was possible to change the
expression

`<t t-out="obj.prop"/>` -> `<t t-out="obj.prop or ''"/>`

While this definitely works, the attribute better signals the intent of
the change.

We thus add support for the attribute on t-out.

closes odoo/enterprise#73004

X-original-commit: c2788e812a06bf9bdabb15f817785e627004a891
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>