[FIX] pos_blackbox_be: allow users to go back on login screen
Behavior before commit:
When opening the POS and selecting a cashier, if the user decides
to go back without selecting any cashier, the clock-in function
gets called anyways.
This leads to pos_blacbox trying to clock in an undefined cashier,
which in turn leads to more errors down the line (including an
unusable pos).

With the modification proposed in this commit, we check that a
cashier has indeed been selected and only then proceed with the
clocking in.

opw-4293988
opw-4182434

closes odoo/enterprise#73009

Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>