[FIX] account_accountant: payment journal entry access
Steps to reproduce:
 - Install invoicing enterprise (but not accounting)
 - Register a payment on an invoice
 - open the form view of the payment

When opening the payment form view, if the payment is associated with a journal entry, which can be accessed with the smart button. However, with invoicing enterprise, this button doesn't make sense because the user will only see the "other info" tab of the journal entry. We therefore limit the access to the button only for accounting users.

task-4224553

closes odoo/enterprise#73036

Related: odoo/odoo#182390
Signed-off-by: William André (wan) <wan@odoo.com>