[FIX] appointment: properly render html intro message when no website
Steps to reproduce:

1. Install only appointments without having a website.
2. Create or use already existing appointments and share them.
3. Go to the share URL
4. Here we will see a dropdown and the description text of the
appointment, try to change the value on the dropdown and see the text.

The issue arises because `replaceChildren` treats strings as plain text,
so HTML in `message_intro` is displayed as text rather than rendered as
HTML.

To resolve this, we use DOMParser to convert the `message_intro` string
into HTML elements. This allows `replaceChildren` to insert actual HTML
nodes instead of plain text, preserving the intended formatting.

This issue was introduce in this commit 9e38975.

opw-4250366

closes odoo/enterprise#73021

Signed-off-by: Maruan Aguerdouh Mohtar (magm) <magm@odoo.com>