[FIX] l10n_cl_edi: display consistent currency rate on printed invoice
Steps to reproduce:
- Install l10n_cl
- Switch to a Chilean company (e.g. CL Company)
- Activate a foreign currency (e.g. USD)
- Make sure that the rate of the foreign currency is not 1
  For example:
  * Unit per CLP: 0.001057876419
  * CLP per Unit:945.29
- Create an invoice in USD with 2 lines having different amount:
  * 15.80
  * 15.00
- Save the invoice
- Print the invoice

Issue:
On the printed invoice, a slightly different currency rate is displayed
for each invoice line.

Cause:
The currency rate to display is computed for each invoice line, based
on the following formula: abs(self.balance) / self.price_subtotal
However, the value of balance is rounded to the unit because of the
CLP currency and therefore it generates a different rate when trying
to compute it.

Solution:
Compute the rate from the total values instead of the subtotal of each
line.
The real rate configured on the currency cannot be used because it can
be modified after the creation of the invoice.

opw-4242448

closes odoo/enterprise#73063

X-original-commit: d174e2b37d4d36a2afa0a003fa04bef4e2e9a739
Related: odoo/odoo#185846
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>