[IMP] documents: simplify Company domain removing pin feature
Purpose
=======
Simplify the company folder domain. Now the domain is just the documents owned
by Odoobot, and without a parent folder. The search panel now uses the same domain,
and so the only difference now between the kanban view and the search panel is that
the search panel only shows folders.

Adapt the "Shared With Me" domain to see documents that are not visible on "Company"
(or "My Drive").

Technical
=========
Because the old `is_pinned_folder` is used in access rule, we can not just set it to
False, it needs to reflect the owner_id / folder_id values (it will be cleaned in master).

Creating in Company is still only allowed for documents manager, and we need to update
the owner to not end up on "My Drive".

Task-4293841

Part-of: odoo/enterprise#73037
Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>