[FIX] documents: noupdate records
server actions that use document tag should be noupdate as well
because document tag records are noupdate, and they could be deleted
by clients.
possible steps to reproduce:
1) make a database 17.0
2) delete document tags
3) upgrade the database to 18.0

Document tags and server actions should be recreated in upgrade scripts
when upgrading to 18.0 if they are missing. Setting the noupdate=1,
makes sure that they are not recreated on subsequent updates or upgrades
after reaching 18.0.
odoo/upgrade#6803

closes odoo/enterprise#73108

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>