[IMP] documents{_sign}: prepare workflow rule migration
Workflow rule can be configured to link a document to any model (the user has
to choose the model when executing the action). In order to migrate such
workflow rule, we make the parameter model of the method action_link_to_record
optional.

Also to ease the conversion of workflow rule into server action, we make the
folder_id optional on the method document_sign_create_sign_template_x of
documents_sign.

Task-4283330

closes odoo/enterprise#73376

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>