[FIX] l10n_ar_edi: Factura B for partner without country
Description of the issue/feature this PR addresses:
Argentinean Localization: Electronic customer invoice Factura B for a customer "Consumidor Final" without country, with "DNI" identification type, with vat and "Consumidor final" afip responsibility type must report to afip the customer vat when the invoice has an amount higher than $344487 is validated but because the vat is not reported to afip then it is not allowed to validate the invoice.

The bug was introduced on this pr:
https://github.com/odoo/enterprise/pull/71562 --> The goal of this pr was to be able to create Factura B for a foreign customer. But prior to this pr the user was allowed to validate an invoice Factura B to a customer "Consumidor Final" without a country set on that customer, with "DNI" identification type, with a vat and "Consumidor final" afip responsibility type when the invoice has an amount higher than $344487

Video explaining the bug:
https://drive.google.com/file/d/1Qb2oUtT26twjCI-pB6oMGMC9gZ6_EBSz/view

Steps to reproduce:

1) Log ing with admin user on runbot odoo enterprise 16 or 17 instance, activate developer mode and install l10n_ar_edi module.

2) Take position on company "Responsable Inscripto".

3) Create an electronic invoice "Factura B" for customer "Consumidor Final Anónimo" with an invoice line with quantity 1 and price 500000. Select electronic journal. The Partner doesn`t have country and has "dni" identification type, dni and "Consumidor final" afip responsibility type.

4) Validate the invoice and then you will receive this message:
Invalid Operation

We couldn't validate the document "Consumidor Final Anónimo" (Draft Invoice *921) in AFIPThis is what we get:

AFIP Validation Observation:
* Code 10015: Para facturas B (CbteDesde igual a CbteHasta) mayor o igual a $344488, DocTipo debe ser uno de los valores devueltos por el mÃ©todo FEParamGetTiposDoc distinto a 99 y DocNro deberÃ¡ ser mayor a 0.

HINT:

 * * If you are making an invoice < 10.000 to an final consumer who has DNI as identification type you must provide the DNI number.
 * Another alternative is to set Sigd (Unidentified / daily global sale) Identification type in order to do not need to inform the Document number,
 * This also happens when you try to generate an invoice with amount greater than 1.000 to an Anonymous Final Consumer instead of one identified with DNI or CUIT.

Edit the invoice and change the partner from "Final Consumer" to a customer with First and Last Name, document number with format 00xxxxxxxx0, where xxxxxxxx is a an eight-digit document number

Please make the required corrections and try again

Current behavior before PR:
It is not allowed to validate Electronic customer invoice Factura B for a customer "Consumidor Final" without country, with "DNI" identification type, with vat and "Consumidor final" afip responsibility type when the invoice has an amount higher than $344487.

Desired behavior after PR is merged:
It is allowed to validate Electronic customer invoice Factura B for a customer "Consumidor Final" without country, with "DNI" identification type, with vat and "Consumidor final" afip responsibility type when the invoice has an amount higher than $344487.

Ticket Adhoc side: 82498
Task latam side: 1283

closes odoo/enterprise#73885

X-original-commit: 7a643a9f3af1c9d0c84915317f358f72f90d3cd0
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>