[FIX] l10n_br_edi: more reliably detect cancellation errors
Before, we relied on just _l10n_br_get_error_from_response() which
checks for the presence of an "error" key in the
response. Unfortunately that only seems to catch errors directly
raised by Avalara. The government can reject the cancellation for a
myriad of reasons [1]. We could hardcode all successful status
codes (24 codes), but to be more robust in case the codes change we
just look if any XML is returned. The lack of XML response should
reliably indicate that the cancellation failed.

[1] 4.4. Lista das Regras de Validação in
https://www.nfe.fazenda.gov.br/portal/exibirArquivo.aspx?conteudo=J%20I%20v4eN00E=

closes odoo/enterprise#73342

Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>