[FIX] web_studio: handle logic for *not* groups attribute
Have a node that goes:
```xml
<div groups="!module.some_group" />

```

Before this commit, this case was not handled and that not-group was just ignored.

This created confusion as the node, when the user was part of the group, would be invisible, the checkbox invisible
would be ticked, and uniticking it had no effect.

After this commit, we handle the case, with a second selector in the studio sidebar.

opw-4273853

closes odoo/enterprise#73419

X-original-commit: c44e5afd65d17f60ae2dad28eeb0054b3536741e
Signed-off-by: Bruno Boi (boi) <boi@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>