[FIX] account_iso20022: Don't report 'SEPA' as Service Level in ISO20022 files
'SEPA' used to be the default for SvcLvl files, which meant even non-SEPA files would have it set to 'SEPA', unless they explicitly overrode the function defining the value of the node. This commit fixes this, by not setting any value to SvcLvl by default, and overriding for SEPA payments.

Part-of: odoo/enterprise#73243
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>