[IMP] account_iso20022: Don't put 'SCT' in InstrId anymore
Before this commit, each payment's InstrId always contained 'SCT' (unless a memo was set on the payment), standing for "SEPA Credit Transfer". For non-SEPA ISO20022, this did not make sense. Since this having this additional info in the field has not additional informative value anyway, we simply remove it.

Part-of: odoo/enterprise#73243
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>