[FIX] account: fix empty string and False/None inconsistencies
When empty, the name field of account.move.line could inconsistently hold either an empty string or False. This inconsistency led to different results when searching for lines with an empty label, as label->not set and label == "" yielded different outcomes. The latter behavior aligns with expected ORM operations.

This fix ensures the name field is set to False whenever the computed result is an empty string, standardizing search behavior. We also adapt some tests to the new behavior in enterprise.

opw-4167139

closes odoo/enterprise#73766

X-original-commit: 52d5c5b4c3867e0d735e84a41fcc1b644fb9ad9d
Related: odoo/odoo#187135
Signed-off-by: William André (wan) <wan@odoo.com>