[FIX] account: fix empty string and False/None inconsistencies
When empty, the name field of account.move.line could inconsistently hold either an empty string or False. This inconsistency led to different results when searching for lines with an empty label, as label->not set and label == "" yielded different outcomes. The latter behavior aligns with expected ORM operations.

This fix ensures the name field is set to False whenever the computed result is an empty string, standardizing search behavior. We also adapt some tests to the new behavior in enterprise.

opw-4167139

closes odoo/enterprise#75834

X-original-commit: f065e604002b3975698a884b3fc42343ddc1b3a7
Related: odoo/odoo#191026
Signed-off-by: William André (wan) <wan@odoo.com>