[FIX] pos_blackbox_be: improve rounding methods UI strings
The rounding method 'HALF-UP' is confusing for regular users when they
see it in the UI. Replacing the user-facing term by the less specific
'Nearest' is clearer, since people naturally expect that to round halves
up.

This commit adapts an error message to the new term.

opw-4266444

closes odoo/enterprise#73316

X-original-commit: 261181bd5ab84418d51310ba7bfa52061ffc893e
Related: odoo/odoo#186279
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>