[FIX] l10n_co_dian: fix wrong usage of countries in manifest
The "countries" attribute in manifest will actually prevent installing
the module when misused. The condition to install the module is an "AND",
therefore you'll need to have both a company pre-existing in the country
AND the dependcy module in auto_install.

Example of right way to use it: auto install l10n_be:
"countries": ['BE'],
"auto_install": ["account"],

Wrong way to use it: auto install l10n_be_reports
"countries": ['BE'],
"auto_install": ['l10n_be']

task-no

Part-of: odoo/enterprise#73315
Signed-off-by: William André (wan) <wan@odoo.com>