[FIX] documents(_project): do not allow admin to see all documents by default
Purpose
=======
Create a new group, so admin don't see all document by default.

Copy in SUDO the folder when we copy a project, to avoid access issue (in case the
project manager has not enough rights to copy the folder). It wasn't catched before
because the test user was erp manager.

Adjust a test in documents_spreadsheet.

Task-4313355

closes odoo/enterprise#73372

Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>