[FIX] account_sepa_direct_debit module description
The module description of account_sepa_direct_debit in the manifest
states that posting an invoice will automatically generate a payment.
Actually that feature was dropped in Odoo 13 with commit
e3d390c4455c620793c481874b4503fa91bb6125 but the `__manifest__.py` was
not updated at that time. This makes the situation uncomfortable where a
feature not present in the code is still advertised in the module
description.

This commit updates the description of the module in the manifest.

closes odoo/enterprise#74532

X-original-commit: 1bb5e6b43ced78a2aa29d45515a4584858a581cd
Related: odoo/odoo#188632
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>