[FIX] quality_control: require quality check only for received products
**Steps to reproduce the bug:**
- Create a tracked product by SN: “P1.”
- Create a storable product: "P2."
- Create a receipt for one unit of P1 and P2.
- Mark it as "To Do."
- Set the done quantity of P1 to 0.
- Attempt to perform the quality check.

**Problem:**
You must perform a quality check for P1 even though it has yet to be
received.

Since the move_line is set to 0 rather than deleted in the original
picking, it will be linked to the backorder when it is created.
Therefore, to avoid unexpected behavior we can ignore linking its
quality check to the backorder.

https://github.com/odoo/odoo/blob/00057f8efd3c728e13a981290a725878e058d2c6/addons/stock/models/stock_picking.py#L832

opw-4187521

closes odoo/enterprise#73407

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>