[FIX] partner_commission: Correcting line description with no subscription end/start dates
Version: 16.0+

Issue:
When we make commission off a subscription sale order the subscription end and start dates are not
passed through the order lines to the invoice lines and we run into a error when trying to create the
corresponding purchase order.

Purpose of this PR:
If there are no subscription dates on the invoice lines then we do not calculate the length of the
subscription in months and instead create the purchase order line description just as the commission
for the corresponding subscription.

Steps to reproduce on runbot:
install partner_commission
set up a subscription with a referrer and a commission plan
create invoice for the subscription
try to register the payment for the invoice
error raised regarding a max function with no args

Notes:
looking to investigate furthe why the subscription dates are not passed from the
subscription to the order lines to the invoice lines

in 17.0+ field name is changed from `subscription_end_date` to `deferred_end_date` (same for `start` date)
will have to adjust in fw ports

opw-4240937

closes odoo/enterprise#73571

X-original-commit: 7b8699706316141ff10b47112b38345aa4237e42
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Ian Dao (iada) <iada@odoo.com>