[FIX] documents: restore scrollbar after preview
Purpose
=======
Fix the view scrollbar which isn't restored when the file preview is closed.

Specifications
==============
When a file is previewed, the view scrollbar is hidden so that
scrolling actions don't impact the view behind the file preview.
However when the preview is closed, the scrollbar isn't restored.
=> Calling the documentList onDeleteCallback method to correctly restore
the view and the scrollbar when closing the preview.

Task-4296803

closes odoo/enterprise#73452

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>