[FIX] pos_settle_due: make total_due visible when searching for more customers
Currently, simple pos users can see the due amount of the customers, unless
the customer wasn't loaded and we need to search for a customer that isn't
loaded.

Steps to reproduce:
-------------------
* Create a new custome, make a sale order and invoice it. This will set
an amount due for that customer
* Connect to the shop with a user that has the `point_of_sale.group_pos_user`
but not the `account.group_account_readonly` group
* Open customer list
> Observation: You can see the amount due for multiple customers
* Search for the customer created
* Select search more
> You can't see the amount due for that customer

Why the fix:
------------
Since simple pos users are allowed to see the amount due for the loaded
customer, there is no reason they shouldn't see it as well for a customer
they need to load.

Commit allowing simple pos users to see (and settle) customer accounts:
https://github.com/odoo/enterprise/commit/37fa4d5f4ed7c7d77f73395a53b7b3ab7006afc4

When loading the pos session `_loader_params_res_partner` is called first
and later is called `_get_pos_ui_res_partner`. It is in the function
`_get_pos_ui_res_partner` that the amount due is compted if the user does
not belong to the group `account.group_account_readonly`.

https://github.com/odoo/enterprise/blob/a35a4755cdf86bcfeda0aca6c06397748ba27362/pos_settle_due/models/pos_session.py#L22-L28

However when we load a customer that wasn't previously loaded, only
`_loader_params_res_partner` is loaded. And since the users does not belong
to the group `account.group_account_readonly`, the field witll not get loaded.

We now compute the amount due with the same logic as in `_get_pos_ui_res_partner`.

opw-4141955

closes odoo/enterprise#73671

Signed-off-by: David Monnom (moda) <moda@odoo.com>