[FIX] pos_settle_due: make total_due visible when searching for more customers
Currently, simple pos users can see the due amount of the customers, unless
the customer wasn't loaded and we need to search for a customer that isn't
loaded.

Steps to reproduce:
-------------------
* Create a new custome, make a sale order and invoice it. This will set
an amount due for that customer
* Connect to the shop with a user that has the `point_of_sale.group_pos_user`
but not the `account.group_account_readonly` group
* Open customer list
> Observation: You can see the amount due for multiple customers
* Search for the customer created
* Select search more
> You can't see the amount due for that customer

Why the fix:
------------
Since simple pos users are allowed to see the amount due for the loaded
customer, there is no reason they shouldn't see it as well for a customer
they need to load.

Commit allowing simple pos users to see (and settle) customer accounts:
https://github.com/odoo/enterprise/commit/37fa4d5f4ed7c7d77f73395a53b7b3ab7006afc4

First we forward the above mentionned commit as it was not fully forwarded
toward master.

With the use of the new relational model in saas-17.2 we cannot use the same
logic as before, i.e. we cannot simply send the value of the customer account
when loading the session. Why? In the case of a simple pos user, you are able
to send values to the session but you cannot send the field because of user
rights. By not sending the field, the relational model does not expect to set
any value for it, thus the values sent will not be used.

what we do frow now is to make an rpc call after the store has been initialized,
the relational models will be in place already, to compute the due amount of all
customer loaded and set them. We also call this rpc after searching for customers
that weren't previously loaded.

opw-4141955

closes odoo/enterprise#77604

X-original-commit: e52fbb7986e1eb918db76d36d7e4a3ab0ac29cba
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>