[FIX] helpdesk_sale_timesheet, project_timesheet_forecast_sale: prevent premature task so confirmation
Steps to reproduce:
- Sales app > Configuration > Settings > Enable 'Quotation Templates'
- Default template > Search more > New
- Set 'Confirmation Mail' to 'Sales: Order Confirmation'
- Project app > New > Enable 'Billable' in the project's settings
- New Task > Set any Customer
- Sales Order Item > Search More > New > Add product then save
- Click 'Sales Order' button

Coupled with https://github.com/odoo/odoo/pull/183586

You'll see in the chatter that the confirmation email is sent before the
product is added to the sales order, resulting in a $0 quotation
document being sent to the customer. This happens because we confirm the
SO at the same time as we create it (After pressing the New button).
Instead we would rather confirm the SO on save, after the order line
creation.

The SO confirmation is necessary because we don't want people to work on
tasks/projects that are still in quotation since they could
hypothetically still be cancelled.

opw-4190402

closes odoo/enterprise#73492

Related: odoo/odoo#183586
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>