[REV] account_batch_payment: add missing dependency to constraint
This reverts commit 4d548d2240c85eebe004595cd65e0fbb6a76a39b.

To reproduce the issue:
1. Create a customer payment and post it
2. Create a customer batch payment
3. Asociate that payment to the batch and validate it
4. Reset payment to draft
====> A constraint on the batch is triggered.

In 17.4, we solved it by removing that check from the constraint, since the same check was done at batch validation: 3e307ec2fa5191fcc657316a125dae208975f82b . In lower versions, the validation does not do that, so we can't apply the same fix. Since the dependency issue can only be annoying in case of customizations, we make the choice to live with it for old versions.

closes odoo/enterprise#73759

X-original-commit: 5d258765155d099ba37a78f5b989081eadb76212
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>