[FIX] account_reports_cash_basis: handle move with multiple receivable accounts
Steps to reproduce:
- Install Accounting and Point of Sale
- Open a "Point of Sale" session
- Create a POS order and pay it by bank
- Create another POS order and pay it by Customer Account
- Close POS session
- Go to Accounting
- Open Bank journal
- Create a statement line and reconcile it with the POS payment
- Go to "Accounting / Reporting / Audit Reports / General Ledger"
- Check the amount of the POS orders in the following journals:
  * [101300] Account Receivable (PoS)
  * [121000] Account Receivable
  * [400000] Product Sales
- Check "Cash Basis Method" in the options
- Check again the amount of the POS orders in the same journals

Issue 1:
The amount for "[121000] Account Receivable" is present but it
shouldn't because no payment has been done for it yet.

Extra steps:
- Create another statement line and reconcile it with the POS
customer account payment
- Check again the amount of the POS orders in the same journals

Issue 2:
The amount of the POS orders is double the original amount.

Issue 3:
If a partial payment is made for one line, the report has
incorrect values.

Cause:
To retrieve the cash basis lines, a complex SQL query is executed.

Let's take the following entries:

- Invoice:

        account        | debit | credit
 ---------------------------------------
  Account Receivable 1 |  100  |   0
  Account Receivable 2 |  200  |   0
  Product Sales        |   0   |  300

- Payments:

        account        | debit | credit
 ---------------------------------------
  Bank                 |   50  |    0
  Account Receivable 1 |   0   |   50

        account        | debit | credit
 ---------------------------------------
  Bank                 |  200  |   0
  Account Receivable 2 |   0   |  200

A SELECT is executed to compute and get the percentage of the paid
amount of each receivable/payable account by move, generating a
temporary "table" as followed:

 move_id |             matched_percentage
 ---------------------------------------------------
    1    | 0.5 (50% paid for Account Receivable 1)
    1    | 1.0 (100% paid for Account Receivable 2)

This table is then joined to account_move_line table on move_id
where matched_percentage is applied to the amounts.

Issues 1 and 2 are due to the fact that it was assumed that each
account move would only contain 1 receivable/payable account,
which is not the case here.
As account_id is not present in the temporary table, it is not
possible to identify to which account a line of that table
refers to.
As there is only a JOIN ON move_id when joining with
account_move_line table, each aml from a move is taken into
account even when it shouldn't.
In this example, they are taken into account twice because the
2 entries have the same move_id.

Issue 3 is coming from the fact that the WHOLE amount of
"Product Sales" is computed with matched_percentage each time.
For example, these 2 lines are computed from the tempory table
for "Product Sales":

* Product Sales = 300 * 0.5 = 150 (for Account Receivable 1)
* Product Sales = 300 * 1.0 = 300 (for Account Receivable 2)

The amount for "Product Sales" sums to 450, which is not correct.
It should be 250.

Solution for issues 1 and 2:
Get account_id and join account_move_line table on it.
An exception should be done for "Product Sales" when joining
account_move_line table because it is not a receivable/payable
account and the temporay table only contains receivable/payable
accounts.

Solution for issue 3:
Compute a ratio for "Product Sales" for each receivable account.
Its amount is 100 for "Account Receivable 1" and 200 for "Account
Receivable 2"
So its ratio should be:

* 100 / 300 = 0.33 (for Account Receivable 1)
* 100 / 300 = 0.66 (for Account Receivable 2)

By applying this ratio, the correct amounts are computed for
"Product Sales":

* Product Sales = 300 * 0.5 * 0.33 = 50 (for Account Receivable 1)
* Product Sales = 300 * 1.0 * 0.66 = 200 (for Account Receivable 2)

opw-4224136

closes odoo/enterprise#74296

X-original-commit: 2dc951a7bbc76c6ae2759cd1bd09b4c3099125cc
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>