[FIX] hr_appraisal: fix feedback section overlap in appraisal form
Steps:
- Install the hr_appraisal module
- Open the employee appraisal form view

Description of the issue/feature this PR addresses:
In the appraisal form view, when feedback is unpublished, the feedback section
overlaps with the chatter area.

Cause:
This issue arises due to unhandled overflow in cases where feedback is
unpublished.

Fix:
This PR resolves the overlap by applying CSS overflow hidden, which hides the
overflowing content and prevents overlap with the chatter section.

task-4320069

closes odoo/enterprise#73645

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>