[FIX] helpdesk_sale_timesheet: avoid using label inside xpath selector
Before this commit, the following selector could crash when the language
is different than english:

`//th[contains(text(), 'Task')]" position="replace">`

Indeed, because `Task` will be translated and so `Task` will no longer
find as text.

This commit replaces that selector to make sure we no longer use a
translated term in the xpath selector.

closes odoo/enterprise#73616

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>