[FIX] account_reports: tax report to be sent activity
The aim of this commit is to allow detection of tax report that still
need to be send easily by putting back the tax report to be sent
activity that was deleted by commit 7bace07feea5

In order to keep things consistent for the end user, the payment
activity will now be a followup activity that can be complete once the
tax report sending task has been complete.

As this is introduced in stable, those activities might not exist an
could require to be created on the fly.

Those creation are planned to be dropped in master.

Context:
The deletion of that activity makes this detection much harder and
removes the possibility to know if was sent or not.

Before the commit:
It is impossible to know if a tax report has been sent or not.

After the commit:
The tax report to be sent can be found easily by looking for the tax
report to be sent activity.
We can assume the tax report was sent once the activity has been
deleted.

closes odoo/enterprise#73615

closes odoo/enterprise#74185

Task-id: 4292304
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>