[FIX] account_reports: fix isNextLineChild
Currently, `isNextLineChild` method will consider
`~account.report~14|~res.partner~10|0~account.move.line~32` being a
a child of
`~account.report~14|~res.partner~10|0~account.move.line~3` as the first
string starts with the second one.
This is wrong, they are siblings.

The fix here is to add a pipe, as done in `isLineChildOf`.

task-4321032

closes odoo/enterprise#77590

X-original-commit: 2695c1eb6e98ac13a7201f510da07ec644103424
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Arnaud Sibille (arsi) <arsi@odoo.com>