[FIX] l10n_mx_reports: filter out root account groups
Steps to reproduce:
- Install l10n_mx_reports
- Switch to a Mexican company (e.g. ESCUELA KEMPER URGATE)
- Go to "Accounting / Configuration / Accounting / Chart of Accounts"
- Create an account:
  * Account Name: [any]
  * Code: 123456789
  * Type: Bank and Cash
- Go to "Accounting / Reporting / Audit Reports / Trial Balance"
- Download "COA SAT (XML)"
- Validate the XML on an online SAT document validator

Issue:
The validation fails because the XML contains lines with incorrect
or missing value for "CodAgrup" attribute.

Cause:
The "CodAgrup" in the "COA SAT (XML)" refers to the code of the
account groups.
The accepted values are defined in the "Catálogo de Códigos
Agrupadores" XSD file.
The created account [123456789] is put in the root account group
with code "1" and a line is added in the "COA SAT (XML)" with this
value.
However, it is not an accepted value.

Solution:
From the account groups created automatically by MX localization,
only the root account groups (i.e. with code "1", "2", "3",...)
do not have a valid code for the "COA SAT (XML)".
These ones can be ignored.
This solution is not perfect as it is still possible to create
an account group with an invalid code that is not a root account
group.
However, handling this use case would require to check that each
code is included in the set of valid codes (there is more than a
thousand).

opw-4209089

closes odoo/enterprise#73866

X-original-commit: 26c4c9c74fef0ed197d7ba982dfb60452a6ee086
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>