[FIX] hr_expense_extract: check if vals['currency_id'] exist in extract
Steps to reproduce:
- Set a main currency and a second one.
- Upload a document in the expense module for the second currency
- Refresh

Cause:
In the for loop there are more than one possible currency detected so `vals['currency_id']` does not exist but the if statement tries to read this value causing an error.

Solution:
Check if the currency_id is in the vals dictionary. If not, the default currency value will be in the Expense.

opw-4307845

closes odoo/enterprise#73660

Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>