[FIX] knowledge: use subroots instead of Apps for Behaviors
Context:
Every Behavior in Knowledge was a new App based on the config of the main Odoo
App.

Among shared resources are registries and services. In particular, the
`main_components` registry is used to mount components, and, for the purpose of
the following example that registry is used by the `popover_service`.

Another thing to note is that the template compilation involves a reference to
the App, and all ComponentNode have an app property which is the app that was
used to compile its template.

Issue:
Now all pieces together in a problematic example case:
Creating a new popover from a Behavior App involves the `main_components`
registry:
- Create a PopoverController (position logic (wrapper)).
  It is created within the main App through the registry, and its lifecycle is
  managed by the main App Scheduler.
- Fill it with a Custom Component (business logic).
  It is created within the Embedded Component App, and its lifecycle is managed
  by the Embedded Component App Scheduler.

Both schedulers lifecycle handling are not synchronized, and at some
indeterministic point one of the Apps will crash during the manipulation
of that popover.

Solution:
Use the new "subroots" OWL feature instead of using sub-apps, so that all
templates are created from the same App, and the scheduler is the same for all
components. This also has the advantage of not having to re-compile all
templates for every Behavior.

Comments fix:
The timing changes introduced by mounting OWL subroots instead of App
highlighted an issue in the comments_handler: comments in the DOM were filtered
in a `useEffect` before executing the preparatory calculations in the
`useRecordObserver` (race condition). The filter is moved in the
`useRecordObserver` callback to avoid that issue. Comments are also not filtered
anymore when switching between handler/panel modes, since the article didn't
change, the comments are the same.

task-4300215

closes odoo/enterprise#73491

X-original-commit: https://github.com/odoo/enterprise/commit/7d676f9406450391347ecdb4e16050847f700313
Signed-off-by: Géry Debongnie <ged@odoo.com>
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>