[FIX] l10n_mx_edi:  ensure CFDI XML document is encoded properly
When uploading a CFDI XML invoice containing the `InformacionGlobal`
element, the system fails to parse the document correctly and displays
an error message.

Steps to Reproduce
------------------
1. Install the `l10n_mx_edi` module.
2. Create a blank invoice.
3. Add a CFDI XML file containing an `InformacionGlobal` as an
   attachment to the invoice. (You can use the file
   `enterprise/l10n_mx_edi/tests/test_files/test_global_invoice_year_month_format.xml`
   for testing.)

The chatter displays the following error:
```
Error importing attachment 'xml_file_name.xml' as invoice (decoder=_l10n_mx_edi_import_cfdi_invoice)
```

Cause
-----
The `InformacionGlobal` element in the XML requires an `Año` attribute
(note the special character `ñ`). Correct encoding is crucial here to
handle special characters accurately. When the XML file is first
uploaded, it is parsed with the correct encoding. However, during
processing, the XML is converted back to a string, losing its original
encoding in the process. This causes issues with special characters like
`ñ`, which can no longer be read correctly by the system, resulting in a
parsing error.

opw-4226968

closes odoo/enterprise#73953

X-original-commit: 82e42c9e1e141d3b8d47a7b9244959826d5db1de
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Séna Serge Nshimiyimana (sesn) <sesn@odoo.com>