[FIX] WhatsApp: WhatsApp Line Break Issue via chats
Before this commit: When a user in Odoo uses the dialog box to send a message to a WhatsApp number and inserts a line break with "Shift+Enter," it creates a <br> tag.
In the WhatsApp module, these tags are stripped, causing the word before and the word after the <br> tag to be joined together without a space or a line break to separate them.

How to reproduce:

Send a message via the dialog box to a WhatsApp number and insert a line break using Shift+Enter.
On the phone, the line break will not be present anymore because the tag <br> is removed with the other tags.

After this commit: the line break is there and the words are not join together anymore.

OPW-4079729

closes odoo/enterprise#73727

Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>