[FIX] hr_gantt: out of bound records in gantt
Steps to reproduce
==================

- Install hr_attendance_gantt,web_studio
- Go to Attendences
- Open studio
- Enable "Display total row" in the sidebar
- Exit studio
- Create a new entry with no end date
- Switch to the week view
- Go to the previous week

=> `Cannot read properties of undefined (reading 'pills')`

Cause of the issue
==================

When records don't have a dateStop value, it is set as now().
The domain included record that start before now and have no stop (so
converted to now).

If the current window is between the 01 and 07 Nov, and today is 14 Nov,
a record starting the 09 Nov that has no end date will be included.

When finding the closest date, it will be out of bound (day 8 when we
only have 7).

Solution
========

Those records shouldn't be included at all as the are outside the
current viewable range.

opw-4160977

closes odoo/enterprise#73839

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>