[FIX] project_account_budget: fix the issue of budget domain
Before this commit:
Clicking on a budgetary position opened the list view showing all budget lines,
regardless of the selected budgetary position.

After this commit:
Now, clicking on a budgetary position will open the list view filtered to show
only the budget lines associated with that specific budgetary position.

Technical:
In `doActionButton`, we use the args if the action type is object,
so `domain` is irrelevant in this context.

task-4195738

closes odoo/enterprise#73785

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>