[FIX] hr_work_entry_contract_enterprise: add year scale
Steps
- install hr_work_entry_contract_enterprise
- Payroll > Work Entries > Select "This year" filter
- ** UncaughtPromiseError > OwlError **

Cause
With the addition of variable time ranges in Gantt, a new
`year` scale is selectable but is missing from the view.

Change
Add year scale. When clicking on a cell or clicking on the
"New" button, the start and end date of the work entries
to create will be the beginning and end of the month.

opw-4321645

closes odoo/enterprise#74593

X-original-commit: a28746ba3d4c95327df4bc990eaafb530bf42b39
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>