[FIX] account_reports: report builder line sequences
Solves 2 issues related to the report builders.
Also cleans the onWillUpdateProps as advised by the Framework JS team.

Issue 1:
Steps to reproduce:
 - Go to Accounting -> Configuration -> Accounting Reports
 - Create a new report with 2 lines
 - drag the bottom line on top of the first one

 -> The list order is properly handled

 - Press the save button

 -> The list order is back to its original form.

Cause of the issue:
When saving, the list.props.list.records is ordered based on its original sequence. This is
due because the "specification" parameter in webSave doesn't contain the proper order of the
line_ids of the accounting report. The reason why it is not present by default is because the
field line_ids doesn't have the "defaultOrderBy" attribute, most probably because the super
function isn't called on the listRenderer.

Issue 2:  -> solved in related community PR
 - Create a report with 2 lines, line 2 being a child of line 1
 - Drag line 2 above line 1

 -> Validation error

Cause of the issue:
 - The backend constraint receives line not ordered by sequence

 Issue 3:
Additionally, this update includes a minor adjustment to the default hierarchy level.
Previously, when creating a new line, the displayed level was 1, but upon saving, it would
unexpectedly switch to 0. This commit resolves that issue, ensuring consistency
in the hierarchy level.

task-4328098

closes odoo/enterprise#78218

X-original-commit: f9c67da8a18e41dc6b18aabff0b7121fc68f3ed1
Related: odoo/odoo#195851
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>