[FIX] documents: prevent MissingError while delete folder and document
Currently, an exception is generated when the user archives a document and its
folder and tries to delete both folder and document once from Trash.

Steps to produce:
- Open Documents > create folder and add document in folder
- Now delete document and folder
- Go to Trash > open list view > Select all records > Click on Delete

error
```
MissingError: Record does not exist or has been deleted.
(Record: documents.document(1,), User: 1)
```

This is because 'to_delete' contains all documents (self), including the folder
that we deleted, and `removable_parent_folders` contains only the folder that
we deleted. Now when cleaning records, [1] unlinks all files, including folders,
and [2] again tries to delete the folder that we already deleted with [1].

This commit will fix the above issue by filtering `removable_parent_folders`
now it is filtering records folders thats id not in self ids.

[1] - https://github.com/odoo/enterprise/blob/5ba1aa66ab17d686da648025c8921995938115ed/documents/models/documents_document.py#L1935
[2] - https://github.com/odoo/enterprise/blob/5ba1aa66ab17d686da648025c8921995938115ed/documents/models/documents_document.py#L1941

sentry-6048388238

closes odoo/enterprise#73816

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>