[FIX] l10n_be_reports: 281.50 and exchange diff
The aim of this commit is to prevent a Zero devision error by excluding
irrelevant 'zeroed' exchange difference.

Context:
Some misc entry with a 0 balance and some amount_currency might be
present and taken into account for the computation.
Such moves have an `amount_total_signed` equal to 0 leading to the
zeroDivisionError issue in the query computation.

Before the commit:
When having such entry considered by the query, creating the 325.50 form
leads to a traceback.

After the commit:
Those irrelevant moves are ignored and the process carry on smoothly.

Task-id: None - Issue spotted on next
X-original-commit: 230134c81fce05ac81a360ad7113af672b320baa
Part-of: odoo/enterprise#74137
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>