[FIX] spreadsheet_edition: reduce thumbnail size
This commit drastically reduces the spreadsheet thumbnails size.

Go to CRM lead and insert the pivot in a spreadsheet:

The size of the thumbnail:

before:		   107Kb
after quality=0.5: 7.8Kb
after quality=0:   2.9Kb

Image quality is indeed reduced, but the thumbnails are always
displayed so small that it doesn't really make any difference
to the naked eye.

Note: I'm not backporting this fix to 16.0 because the code changed
and webp was not supported at the time (even though we could use
jpeg)

closes odoo/enterprise#74295

Task: 4337496
X-original-commit: 22186b61b15bbbd64e8e99f131c5abeccf750b56
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>