[FIX] stock_barcode: dont split move if barcode was not mounted
Steps to reproduce:

- Create a delivery order for 2 units of a product
- Scan 1 unit and click on the pencil icon to edit the record further
- On the digipad access the backend by clicking on the prodcut name.
- Come back to the record the faster you can.
> `Error in Owl lifecycle: undefined record in get moveIds`

Cause of the issue:

Since the barcode app relies on stock move line's data to generate its
main component it is necessary to split moves in DB when you exit the
barcode app in order to rerender the main component with appropriate
data the next time you enter the barcode. Currently this split is done
onWillDestroy:
https://github.com/odoo/enterprise/blob/c97daa0127725411ba16ed8c94e6b31a0f4e1cc3/stock_barcode/static/src/components/main.js#L105-L107
https://github.com/odoo/enterprise/blob/c97daa0127725411ba16ed8c94e6b31a0f4e1cc3/stock_barcode/static/src/models/barcode_picking_model.js#L1327-L1329
However, if you come back to the barcode app fast enough you will start
the creation of a new barcode main component before this rpc call has
ended, in turn you will modify the content of the component before it
has even mounted (even prior to the assignment of this.record) and owl
will therefore destroy that component and recreate a new one with the
appropriate content. But since this non-mounted component is destroyed
it triggers the onWillDestroy and since it has no record yet it triggers
an error by trying to access `this.moveIds`:
https://github.com/odoo/enterprise/blob/c97daa0127725411ba16ed8c94e6b31a0f4e1cc3/stock_barcode/static/src/models/barcode_picking_model.js#L656-L658

Fix:

Since we only want to trigger our `_onExit` method to update DB's data
if the component has been mounted and since the `this.record` will
always exist in that case, we should use the onWillUnmount hook rather
than the onWillDestroy.

Follow-up of c16acf3617442d3048f6a1246a2bc826af3c02fd (fixed in forward
ports post 17.0).

opw-4232106

closes odoo/enterprise#73894

Signed-off-by: Steve Van Essche <svs@odoo.com>