[I18N] l10n_co_reports: (re)-add translations
During the es_XX.po cleaning, the po file for l10n_co_reports was
incorrectly not converted to es_419.po and only deleted. We now restore
it + correctly rename it to es_419.po

As part of this cleaning, the pot file has been re-exported and the po
re-synced with it. Missing translations were also added it. A
sentence that was obviously incorrect in the original Spanish was also
translated to a correct sentence.

opw-4336377

closes odoo/enterprise#74248

Signed-off-by: Louis Wicket (wil) <wil@odoo.com>