[FIX] l10n_be_hr_payroll_account: fix querycount
Due to including the employee id in some of the reports generated during
the test, some query count could break when flushing data to the
database -> file size of documents could not be batched properly.
By adding a zfill on the employee's id in the template during the test,
we can ensure the same filesize regardless of the employee's id.

Runbot Error 4650

closes odoo/enterprise#73692

Related: odoo/odoo#186983
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>