[FIX] approvals_purchase: traceback when creating a vendor on the fly.
Steps to reproduce:
  1. Open the `Approvals` app.
  2. Select `Create RFQs`.
  3. Add any product.
  4. Try to create a new vendor for the added product.
  5. An error will occur.

Explanation of the bug:
In order to specify a vendor, one needs to specify
the vendor data, the price of the product and the
quantity of the product sold for that price.

So, when the user starts typing to create a new vendor
and then presses on `Create`, an error occurs because
it can't be determined which field should be filled with
the entered data. i.e. will the entered data be used to fill
the vendor data, the product price or the product
quantity?

This commit fixes the bug by only allowing
`Create and Edit` option for the vendor column.

task-4337135

closes odoo/enterprise#73936

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>