[FIX] pos_order_tracking_display: sync order status over multi displays
There is currently a problem of synchronization when using multiple tracking
displays per pos config.

Steps to reproduce:
-------------------
* Configure 2 Preparation displays, 1 for the category food, the other for
drinks. Link them both to the restaurant
* Open the two preparation displays and their preparation status
* In the restaurant, place an order for 1 food and 1 drink product
* On one of the preparation display mark the food as ready
> Observation: Both status screen will show 'ready'
* Again on the preparation display mark the food as completed
> Observation: The order disappears from both screen.

Refreshing both status screen will not show the same state

Why the fix:
------------
The status screen should be seen as an ORDER status screen.

When placing an order 1 in the restaurant, the status screen should only show
it as ready when all the pos preparation display orders linked to that order
are ready.

This idea behind this fix is to get the status of the order as they are in the
related preparation display and to modify it with regards of other status screen
that are linked to the same pos config.

Let's discuss this part:
```python
        for display in other_displays:
            other_display_orders = display._get_pos_orders()
            for order in other_display_orders['notDone']:
                if order in current_display_orders['done']:
                    current_display_orders['done'].remove(order)
                    current_display_orders['notDone'].append(order)
```
`other_display_orders` represent the preparation display orders from other preparation displays.
`current_display_orders` represents the preparation display orders related to the
preparation display we just used.

Here we basically ensure that we don't show a preparation order on the status
screen as done if there is another preparation order linked to the same order
which is not ready.

Scenario:
* `current_display_orders = {'done': ['101', '102'], 'notDone': ['103']} `
* `other_display_orders = {'done': ['101'], 'notDone': ['102', '103']}`
* result -> `current_display_orders = {'done': ['101'], 'notDone': ['103', '102']} `

The fix does not stop here as we also need to consider 2 scenarios:
* 1, we mark a preparation display order as complete in one preparation display.
* 2, we have preparation display orders that are not on the preparation display
but still linked to the same pos config as the status screen opened

In both scenarios, we will have some traking numbers in `orders` which will
not be found in `current_display_orders`.

```python
            for order in other_display_orders['done'] + other_display_orders['notDone']:
                if (
                    order not in current_display_orders['done'] + current_display_orders['notDone']
                    and self.env["pos_preparation_display.order"].search(
                        [('pos_order_id.tracking_number', 'ilike', order)]
                    ).pos_config_id.id in self.pos_config_ids.ids
                ):
                    stage = 'done' if order in orders['done'] else 'notDone'
                    current_display_orders[stage].append(order)
```

This part of the fix will look at all the orders that are not present in
`current_display_orders` and if the preparation order belong to a pos order
made in the pos config of the current preparation display status then it
will be added in the respective stage.

We also apply this logic when loading the status screen for the first time/
or refreshing it. We want to see the status of all orders from the set of
pos configs.

opw-4265987

closes odoo/enterprise#74450

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>