[FIX] account_{followup,reports}: fix error when sending partner ledger report
Currently, a traceback is occurring when the user tries to send
the partner ledger report after uninstalling account followup.

To reproduce this issue:

1) Install `Accounting`
2) Uninstall `account_followup`
3) Open the partner ledger report from reporting
4) Click the `send` button and then `send & print`  button

Error:-
"'res.partner' object has no attribute '_get_partner_account_report_attachment'"

The `_get_partner_account_report_attachment` method is defined in
`account_followup` and is referenced in `account_reports`.

https://github.com/odoo/enterprise/blob/6100796466f356053dde87110c80812c2dcba3ba/account_followup/models/res_partner.py#L295

When the user uninstalls the `followup` it leads to the above traceback
when the below line of code executes.

https://github.com/odoo/enterprise/blob/6100796466f356053dde87110c80812c2dcba3ba/account_reports/wizard/account_report_send.py#L250

We can resolve this issue by moving the code from `account_followup`
to `account_reports`. As we can see that `followup` depends on `account_reports`

sentry-6069400574

closes odoo/enterprise#74064

Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>