[FIX] account_accountant: broken focus step on new Chrome
Since the "new" Chrome's headless mode (default starting with version
128), the tour testing the accounting's bank reconciliation widget fails
on some steps while selecting and editing the entry's line amount.

More precisely, but without absolute certainty, this looks to be a
concurrency issue between the selection of the line, the update of the
notebook's form and the focus of the related field (for quick edition).

While the failing step relies solely on the field being focused (which
is finicky on headless browser where the "focus" needs to be emulated),
this commit works around this issue by checking on the entry's line
selected instead.

Note: while being an issue in automated test, the feature operated by an
actual user is working properly.

closes odoo/enterprise#77905

X-original-commit: 1f2e34670689c1e1fa0d9db2217c8be26384d68c
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: Pierre Paridans (app) <app@odoo.com>