[FIX] knowledge: update id references when copying an article
In the editor refactoring done in Odoo 18, we slightly changed the
structure of the embedded props stored in the anchor node of the
embedded views. Now, the view context is nested inside the `viewProps`
attribute of the embedded props.

When the user creates a copy of an article using the "Create a Copy"
button, we iterate over all the embedded views stored in the article
and we update the context of the embedded views. This way if they list
the article items of the original article, they must list the article
items of the newly created article. This logic was introduced to allow
people to start from a fresh and "ready-to-use" article.

As the context has been moved in the embedded props, the check verifying
if the embedded view has a context always fails and the references stored
in the context are never updated. As a result, the newly created article
has embedded views listing the article items of the original article
which is not what we want.

To fix the issue, we will read the context from the `viewProps` attribute
of the embedded props and remove the last calls to `urllib` we made to
encode/decode the embedded view props.

Steps to reproduce:
1. Open Knowledge
2. Create an article
3. Insert an embedded view with `/kanban`, `/card` or `/list`
4. Create a few article items from the embedded view
5. Create a copy of the article you created at step 2 by clicking on the
   "Create a Copy" button from the kebab menu.
=> The embedded view in the new article lists the article items of the
   original article and not its own article items.

TO BE: As it was the case before Odoo 18, the embedded views should list
the article of the newly created article.

see: https://github.com/odoo/enterprise/pull/42936
see: https://github.com/odoo/enterprise/pull/67083

task-4343222

closes odoo/enterprise#74044

Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>