[FIX] data_cleaning: prevent error when deleting storage record
When the user tries to delete the storage record,
A traceback will appear.

Steps to reproduce the error:
- Install "data_cleaning" module
- Go to Data Cleaning > Storage > Select any record > Actions > Delete > Delete

Traceback:
"UndefinedTable
relation "ir_attachment_report" does not exist
LINE 1: DELETE FROM "ir_attachment_report" WHERE id IN (1"

"ir.attachment.report" model is "_auto = False".
So, a database table will not be created for the model.
When the user tries to delete the record of that model,
It will lead to the above traceback.

sentry-6073448766

closes odoo/enterprise#74074

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>