[FIX] accountant, account_accountant: Fix accounting tours
Problem
---------
When launching the accounting tours, the don't actually run.
This is because, when accountant is installed, the App menu item
name is different than when only Invoicing is available. Furthermore,
the last steps of the tours where failing because the buttons had been
renamed.

Solution
---------
When accountant is available, the app menu name should be
`accountant.menu_accounting`, this is done using a patch of the
`goToAccountMenu`. Furthemore, the button names for the last steps have
been updated accordingly and a few extra steps have been added to
generate a transaction in the bank journal.

task-4309716

closes odoo/enterprise#74092

Related: odoo/odoo#188073
Signed-off-by: John Laterre (jol) <jol@odoo.com>