[PERF] stock_barcode: speedup loading operations in barcode app
Currently, when an operation is loaded in the barcode app, the frontend
executes a request to get_barcode_data which will perform a read on all
relevant barcode data from the backend to cache it on the frontend. The
problem is, when loading the records of stock.quant.packages, it will
attempt to read all empty packages with no location id. There are cases
where these packages are way too many i.e. 500K and thus to read fields
on these records, it would be terribly slow. As this is done upon
opening every operation in the barcode app, it's making the app feel
sluggish.

To workaround the problem, this PR introduces a config paramter "stock_barcode.usable_packages_limit" which will limit the total number of empty packages being read to feed the frontend cache.
This gives the option to tradeoff between cache locality and network
connection. Setting this limit to a smaller number will result in more
rpc calls instead of suffering a slowdown on one big rpc call to load
all data once.

Benchmarks:

This will totally depend on the configured limit. Lower limit would mean
higher speedup and vice versa. As an example, setting a limit of 1000
cut down the loading of get_barcode_data from 13.87 s to 768 ms

opw-4245980

closes odoo/enterprise#74466

X-original-commit: 65d88de302166301d99e0c991fa43e13d9a16096
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Mohamed Kasem (kasm) <kasm@odoo.com>