[IMP] sale_subscription: fix invoice on churn sub
Before this commit, when the end_date occured inside the next period, we
could invoice the prorated part of the period.
Invoice from 5th of November to 5 of December.
End date on the 1st of December.

The invoice would be from the 5th of November to the 1st of December.
As this seems smart, it would disrupt some flows where salesperson would
put an end_date as an incentive to force people pay their
contract/upsells in a fear of service disruption.

With the previous code, the invoice would be prorated and removing the
end_date would mess up with the next_invocie_date and invoicing periods.

closes odoo/enterprise#74181

Taskid: 4358288
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>