[FIX] documents: do not propagate access change on frozen folders / spreadsheets
Bug
===
A system administrator can view all frozen folders / spreadsheet, because of that,
when he changes the access on a folder, it can be propagated on the frozen folders
and on the frozen spreadsheets inside.

But, the frozen spreadsheets can not be editable. To solve that issue, we now
need to open the share panel of the frozen spreadsheets to change their access.

Fix a bug where the documents access could be created many times for the same partner.

Task-4346980

Part-of: odoo/enterprise#74148
Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>