[FIX] delivery_starshipit: fix TNT manifest issue
If Starshipit is configured with TNT service (Australia), the API does
not return a pdf field for `orders/manifest` endpoint. This causes the
transfer validation process to fail in Odoo. This fix does not resolve
the core problem, but it will allow the stock picking to be marked as
"Done", with tracking number and link available to the user. The
downside is that the manifest PDF will not be available

closes odoo/enterprise#74190

Task: 4195503
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>