[ADD] l10n_ke_hr_payroll_shif: NHIF to SHIF
Problem
----------
Since 1st October, the SHIF replace the NHIF, the only difference is the computation of the amount.
All payslip before and on the 9 October use the NHIF and all new payslip from the 9 October use the SHIF

Objective
----------
Adapt the salary rules to compute the good rule.
Adapt the NHIF Report to make NHIF or SHIF reports.

Solution
----------
Create new SHIF rule parameter with a min amount and rate computation.
Create a new report wizard to manage the 2 different rules NHIF and SHIF

task-4294419

closes odoo/enterprise#74242

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>