[FIX] l10n_cl_edi_pos : show the date on receipt even without VAT
Steps to reproduce:
- Install "l10n_cl_edi_pos"
- Create a new company with the Chilian localization
- Do not add a VAT number
- Go to POS, create an order and validate
- The date is not displayed on the ticket

Cause:
The localization is moving the date from the bottom of the receipt to the header but the conditions are not the same:
- If the country of the company is Chile then date is removed from the bottom
- If the company has a Chilian VAT the date is shown in the header
So there is a case where none of the conditions are True and the date does not show.

Solution:
Add a condition on the VAT to delete the date from the bottom.

opw-4330252

closes odoo/enterprise#74533

X-original-commit: ba75a2ab223350cff1e2b753ef129a97a19ecc38
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>