[FIX] hr_payroll: fix payslip PDF report to reflect schedule pay
Steps:
- Install the hr_payroll module
- Open the payslip form view
- click on print

Description of the issue/feature this PR addresses:
When the schedule pay on a contract is set to anything other than `Monthly`
(e.g., bi-weekly), the PDF report incorrectly displays `Monthly`. This happens
because the report only considers the wage type (monthly or hourly) and does not
take the schedule pay into account.

Fix:
This PR addresses the issue by including the schedule pay type in the PDF report
when the wage type is set to `Monthly`.

task-4332531

closes odoo/enterprise#75342

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>