[IMP] hr_contract_salary: dropdown group
Problem
----------
In th salary configurator, The display "group by category" for dropdown is broken in 18.0 due to framework task.

Objective
----------
Make it works again like in 17.4.
Group options by optgroup
Add space under the dropdown

Solution
----------
add groups in the state and fill it with label of groups (attribute in optgroup) and all the choices (options inside optgroup)

task-4350875

closes odoo/enterprise#74304

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>