[IMP] knowledge: file box style
This commit, alongside changes in html_editor, changes the style of the
file box (result of the /file command).

Formerly a banner, the new file card has a more compact style and no
longer has a download button. The new style allows for grid display of
multiple files in the same line.

task-3522395

Part-of: odoo/enterprise#75045
Related: odoo/odoo#189203
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>