[IMP] documents: propagate the access changes in the trash
Purpose
=======
Propagate the access changes (member changes, access_internal, access_via_link, etc)
in all documents inside a folder, even if we sent them to the trash.

The interface still disables the share button (it's weird to share archived documents,
but it's possible to call `action_update_access_rights` on them in an action eg,
for flexibility).

Task-4353456

closes odoo/enterprise#74355

Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>