[FIX] helpdesk_stock: display products from parent company
Steps to reproduce:
- Install helpdesk_stock and Studio
- Create 3 contacts, 1 company and two child employees
- Create a quotation for each contact with a different product
- Helpdesk app > Create a ticket
- Add the Product ('product_id') field with studio
- (Under 'Existing Fields' search product)
- (View tab > Show invisible elements > Find product field > untick invisible)

The products from contacts linked to the parent company should appear,
but only those related to SOs from the contact itself or its parent
company are shown (i.e. We should see all 3 products no matter which
contact is set as customer).

This is related to https://github.com/odoo/enterprise/pull/73391, which
was a step in the right direction but still insufficient. Unlike that
fix however this one is relevant up to master.

opw-4285382

closes odoo/enterprise#75748

X-original-commit: 0545df39f6d0fe925862134269a77993437f7c40
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Youssef Bashandy (yoba) <yoba@odoo.com>