[FIX] l10n_be_hr_payroll: sickness relapse box visible
A checkbox to indicate if a Sick leave is a relapse of a previous sick leave was added to version 18.0. Regular users can see the checkbox when creating or viewing sick leave while the intention was for only time off officers to be able to see it. This patch fixes this issue by assigning the checkbox to the correct group.

task-4354558

closes odoo/enterprise#74422

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>