[FIX] l10n_mx_edi_landing: add customs emission date to invoice lines
To comply with Mexican tax regulations (Article 29-A, section VII of the Fiscal Code and the CFDI guide, pg. 22), both the customs number and its emission date must be included in product descriptions for national sales.

This commit adds the date(s) of the emission of the customs number(s) to every customer invoice line label

task-4132754

closes odoo/enterprise#75268

X-original-commit: 413a8de7d7636cfd917fcab3f6d202b8b45cffa2
Signed-off-by: Laurent Smet (las) <las@odoo.com>