[REF] *: sanitize run function in tours
In commit, we remove empty runs that no longer have a reason to exist.
We also prefer to use tour helpers ( hover ) instead of using
makeVisible functions in order to simulate user interactions as best as
possible.
We take advantage of this commit to remove the nextTick helper in
webStudio (we no longer need it since 3bfd2e9cf8 )

closes odoo/enterprise#74252

Related: odoo/odoo#187999
Signed-off-by: Bastien Pierre (ipb) <ipb@odoo.com>