[FIX] mrp_workorder: disable mrp dialog buttons while closing
Steps to reproduce:

- Create a bom for a product FP:
    - 1 x COMP (component)
    - operations: OP1:
        - instruction: - type: pass/fail,
- Create and confirm an MO for 1 unit of FP
- Go to the shopfloor, click on the instruction to open a dialog.
- Spam the Pass button

> Traceback: Uncaught Promise:
> Cannot read properties of undefined (reading 'resId')

Cause of the issue:

Clicking on a footer button of the `MrpQualityCheckConfirmationDialog`
will trigger a call of the `doActionAndClose` method:
https://github.com/odoo/enterprise/blob/55c6292ff9d59ff09b1d81a61752765cf70d72c0/mrp_workorder/static/src/mrp_display/dialog/mrp_quality_check_confirmation_dialog.js#L93-L113
A the end of this call a `qualityCheckDone` is called to complete the
operation which to work properly expects a `lastOpenedQualityCheck` to
be set. However, at the end of this call, `displayInstruction`
is called to generate a new MRP dialog for the next instuction if even
exist. And, during this call if th ecurrent record (instruction)
correspond to the `lastOpenedQualityCheck`, the value will be erased:
https://github.com/odoo/enterprise/blob/55c6292ff9d59ff09b1d81a61752765cf70d72c0/mrp_workorder/static/src/mrp_display/mrp_display_record.js#L337-L340
This causes the error because if you click too fast you will clean the
value that you are suppose to use in the next call `qualityCheckDone`
which will explode.

Fix:

We diasble the buttons of the Dialog once a a call of the
`doActionAndClose` is triggered. That way, it becomes impossible to
spam an action that is already processing the info.
This is a backport the part of e4190f1ea2c89439d2cd4afa5137d0fedb06431e
related to our issue. In addition to this bacport we apply this change
on the `MrpRegisterProductionDialog` extending the
`MrpQualityCheckConfirmationDialog` to limit the amount of rpc called
performed by validating a registration (prior, you could spam the
validate button to encode the exact same value multiple times).

Note:

By opposition to e4190f1ea2c89439d2cd4afa5137d0fedb06431e we do not
disable the buttons during the `openWorksheet` since a priori this
action does not close the dialog and we might want to press other
buttons on the dialog once this action has been performed.

opw-4352828

closes odoo/enterprise#74795

Signed-off-by: Steve Van Essche <svs@odoo.com>