[FIX] sale_renting, *: align checkbox with label text in mobile and form view
* = [sale_subscription]

Before this commit:
In the mobile and form view, when a checkbox field had a long label, the
checkbox would appear on top, with the label text displayed below.

After this commit:
In the mobile and form view, checkbox fields with long labels are now properly
aligned, with the checkbox and label text displayed.

Task-4269578

closes odoo/enterprise#74561

Related: odoo/odoo#187980
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>