[FIX] documents: show the "activities button" in chatter
Before this commit:
In the Kanban view, when selecting request documents and opening the chatter,
the activity button does not appear.
Similarly, in the list view, when selecting any documents and opening the chatter,
the activity button is not displayed.

Technical:
For request documents, we get an empty res_model, so a condition was added to
handle this.
In the documents_list_renderer, the 'res_model' cannot be retrieved, since active
fields are calculated from the 'arch info'. Therefore, we added the field in the list view.

After this commit:
The activity button will now be visible for all documents in both the Kanban and list views.

Task-4341409

closes odoo/enterprise#74588

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>