[FIX] account: Prevent exchange diff if rates are same
When creating a write-off for multiple multicurrency journal items that
where created with the same rate, the reconciliation wizard will
recompute the balance of the write-off based on the residual amount of
the journal items in invoice currency and the reconciliation rate.

Because each journal item applied rounding separately, the resulting
amount might differ from the residual amount of the journal items in
company currency, which will trigger the creation of an exchange
difference entry (for a few cents).

This entry is spurious and confusing for the user because the reconciled
journal items all shared the same currency and rate and therefore the
reconciliation does not correspond to a currency exchange event.

After this commit, when computing the residual amount to write-off in
the reconciliation wizard, we check whether the journal items used the
same rate as the reconciliation rate, and in that case we directly take
the residual amount of the journal item in company currency instead of
recomputing the residual amount based on the invoice currency amount and
the reconciliation rate, thus avoiding rounding errors.

task-4247344

closes odoo/enterprise#74570

Signed-off-by: Laurent Smet (las) <las@odoo.com>