[FIX] pos_preparation_display: send cancelled notification
Currently, when cancelling an order, the notification is not sent to the
preparation display.

Steps to reproduce:
-------------------
* Open both restaurant and preparation display
* In the restaurant place an order and send it to the kitchen
* Select Action > Cancel order
> Observation: The cancellation is not sent to the preparation display

Note: If you send a new order or refresh the preparation display then you
can see that the order has been cancelled.

Why the fix:
------------
When cancelling order, data looks like this
`{'change': True, 'sound': False, 'category_ids': {}}`. Since `category_ids`
is empty, `preparation_displays` will alsos be empty which explains why
the preparation display does not receive the notification, it is simply not
sent.
https://github.com/odoo/enterprise/blob/42b7b322d6c720de45661e3370397adaf74fe1f4/pos_preparation_display/models/preparation_display_order.py#L29-L40

Since all lines are canceled, we include all the different categories they
cover. Now `preparation_displays` is not empty and all the displays it covers
will receive a notification.

opw-4322600

closes odoo/enterprise#74683

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>