[FIX] account_consolidation: prevent comparison to be auditable
Steps to reproduce:
[l10n_be_us_consolidation_demo]
- go to consolidation > Test Chart > Consolidated Balance > Comparison > Test Chart
- click on a cell to proceed to an audit

Issue:
Stacktrace

Cause:
Journal id is not defined. When there are multiple periods we use the `ComparisonBuilder`
https://github.com/odoo/enterprise/blob/4367042c8c7f2190d23019e591c332045b7d4993/account_consolidation/report/trial_balance.py#L139

Which, contrarily to the DefautBuilder, journals is not defined.
In the AbstractBuilder (inherited in the two builders), `auditable` is set to True without consideration of the builder's definition
https://github.com/odoo/enterprise/blob/3e5ac3554f4fded90fd37c56c7c60723aec59404/account_consolidation/report/builder/abstract.py#L278

On the front side, whenver we click on a auditable cell, we call the action_open_audit but journal_id is defined
https://github.com/odoo/enterprise/blob/2d774344109c2f9b55d9ea6160ae99aeca4f0420/account_consolidation/static/src/components/consolidation_report/line_cell/line_cell.js#L19-L38

-> error

Solution:
Asusming that `auditable` is the rule and the case for the comparison builder is the exception, auditable is set to False only in the ComparisonBuilder

Since `consolidation` is meant to be decommissioned, the least impacting fix is being applied.

Note:
in 16.0, the cells are not auditable when using comparison periods.
opw-4265031

closes odoo/enterprise#74577

Signed-off-by: William André (wan) <wan@odoo.com>