[FIX] approvals: contact creation from approval
Issue:
When the user tries to create a contact from the approval form, they get a "Missing Record" error.

Steps to reproduce:
- Install the Approvals app
- Create a new Approval Type with Contact set as either 'Required' or 'Optional'
- When creating a new Request from the created Type, try to create a new Contact

Cause:
Both Approvals and Partners have a 'category_id' field.
In the approval form, the context has the approval's value for 'category_id', which is being used for the partner creation.

Solution:
In the approval form view, change the context used by the partner field to not have a default 'category_id'.

Ticket:
4328371

closes odoo/enterprise#74579

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>