[FIX] account_base_import: Don't skip tests if openpyxl installed
Since `xlrd >= 2.0` dropped XLSX support, we use `openpyxl` to open
XLSX files if the `xlrd >= 2.0` is installed.[^1]

However, the tests in `account_base_import` are skipped unless
`xlrd.xlsx` can be imported. As a result, they are not run on runbot,
where `xlrd >= 2.0` is installed.

We therefore need to avoid skipping them if openpyxl is installed.

runbot-108001

[^1]: https://github.com/odoo/odoo/pull/169245

closes odoo/enterprise#74669

X-original-commit: 5e968dfce025102bc27f01ba9b07304477835b47
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>