[FIX] documents: fix document_delete_tour failure
Following odoo/enterprise#73372 where we create a new group, so that the admin
users don't see all the documents by default, the admin user used in this tour
is not able to see anymore the "Folder1" created in the test as it is owned by
odooBot with no additional rights.

To solve the problem, we set access_internal="edit" on the "Folder1".

Task-4361749

closes odoo/enterprise#74614

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>