[FIX] social_linkedin: avoid unnecessary disconnections
This commit fixes an issue with the social_linkedin module's comments fetch
methods introduced in #67079.
This method disconnects the user from its account every time they open a post's
comments, even if all API calls are successful. This shouldn't happen as nothing
wrong happened to warrant a disconnection on our end.

Now the user's account is disconnected if and only if there was an actual issue
when fetching comments.

task-4363737

closes odoo/enterprise#74820

X-original-commit: c754518dd34e8c1904c7ce4f0408b47af658bb9d
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>