[FIX] web_gantt: rangeId not in scales
The attribute scales of the gantt view specifies which scales are available in
the view. This means that for instance "year" or "month" can be absent from
scales. If that happens and that "year" or "month" is the selected range, a crash
occurs when rendering the GanttRendererControls component:
    TypeError: Cannot read properties of undefined (reading 'groupHeaderFormatter')

Here we assign to each range a function groupHeaderFormatter so that we can call
it independently from the available scales.

closes odoo/enterprise#74891

X-original-commit: 0fa5119de1e5763c0b578324b9311af0e8f8064c
Signed-off-by: Bruno Boi (boi) <boi@odoo.com>
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>