[FIX] website_appointment: use recaptchav3
recaptchav3 is enabled on most website forms by default
but not on appointments.

To mitigate bot attacks spamming free appointment bookings
we enable the feature when website is installed.

A similar change was recently applied at [1]

Forward port of [2] in PR odoo/enterprise#74682

[1]: bafa915f85fb8fc6ca2ae7d194a4593cb2463c2e
[2]: 3dc54dae2dbc43f26b3e6c8bf852ac21d5dfff2a

task-4335141

X-original-commit: 474a0afe8bdb76af453687e01453793c895fdac3
Part-of: odoo/enterprise#80613
Related: odoo/odoo#200049
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>
Signed-off-by: Renaud Thiry (reth) <reth@odoo.com>