[FIX] spreadsheet_edition: show name of templates
when creating a new spreadsheet in document, the user is prompted to either choose a blank spreadsheet
or choose to create them from a template. The view didn't list the name of the template.
This commit shows the name of the template

The tour `spreadsheet_create_list_view` that was delete by accident in
https://github.com/odoo-dev/enterprise/commit/a32825ee00f2b330d99113f4d8c1488903fe744e was restored.
We now use the list of apps instead of the list of documents in this tour.

closes odoo/enterprise#75515

Task: 4363974
X-original-commit: f460b38454b2c1ab5900e9a2f1d5c5bdef034302
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>