[IMP] website_helpdesk_knowledge: adapt tour
Due to changes in the behaviour of the /file commmand in html_editor, it
no longer opens the media dialog, but the system's file selector
directly.

This commit adapts the tour to use the /media command instead, which
does open the media dialog.

task-3522395

Part-of: odoo/enterprise#75045
Related: odoo/odoo#189203
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>