[FIX] whatsapp: force unpin of whatsapp conversations after 5 days
Follow-up of https://github.com/odoo/enterprise/pull/74131

PR above attempts to solve an issue of every-growing unpinned
whatsapp conversations that make it harder to use Discuss for
performance reasons, due to reaching the technical limit of
around 2000 pinned conversations. To do so, it automatically
forces unpinning whatsapp conversation after 2 weeks of inactivity,
even if the conversation was unread.

While this is an improvement, if there are around 150 whatsapp new
whatsapp conversations made every day, then the 2 week force unpin
is not enough, because the 2000 pinned conversation limit is reached.

This commit fixes the problem by reducing the force unpin of whatsapp
conversation to 5 days.

The architecture of Discuss should scale better, but this is hardly
achievable in short manner, hence why the current solution to this
problem is prevent reaching a high amount of pinned conversations.

opw-4330847

closes odoo/enterprise#74703

Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>