[FIX] sale_subscription: ensure upsells don't update the next_invoice_date
When an upsell is invoiced, the parent subscription_id value is stored on the
account.move.line. When that invoice is posted, the subscription_id is
used to increment the next invoice date.

When everything is working well, the deferred_end_date of the upsell
invoice is identical to the next_invoice_date of the subscription.

On system with older historical values, these dates can be different.
The system should be more resilient. As an upsell should never update
the next_invoice_date of a contract, we filter them out.

Part-of: odoo/enterprise#74234
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>