[FIX] documents: maintain owner access after pinning
Follow-up of 4d653bf5 where we forgot to make sure that the
owner keeps an `edit` permission on the documents set as company
root.

When moving a folder in Company to pin it, we do not want to add
the member on all children (default behavior).

Tests are adapted.

Task-4365165

Part-of: odoo/enterprise#74713
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>