[FIX] l10n_mx_edi: retry in CFDI on 'Print & Send' Invoice Warning
This error occurs when attempting to 'Print & Send' an invoice using incorrect
credentials for the 'Authorized Certification Provider (PAC)' or if no
credentials are provided and you choose to 'Retry' after receiving the warning.

Steps to reproduce:
- Install 'l10n_mx_edi' module
- Change the company to 'ESCUELA KEMPER URGATE'
- Invoicing > Customers > Invoices > Select an Invoice
- 'Print&Send' (you will face a warning) > CFDI(Notebook page) > Retry

Traceback: TypeError
'int' object is not iterable

At [1], we are directly passing the ID of self instead of self.

[1]- https://github.com/odoo/enterprise/blob/0f7a4eb991fc59240710c0f8ba78f4da5b889fe6/l10n_mx_edi/models/account_move.py#L1726

sentry-6095574663

closes odoo/enterprise#74758

Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>