[FIX] documents_hr_contract: fix sign contract
while signing a contract, it may happen that we end up with a sign request
linked to multiple contracts. This will lead to an error when trying to
call `_get_sign_request_folder` that has `ensure_one` on the contract.
to avoid that we take the newest contract linked to the sign request.

opw-4348211

closes odoo/enterprise#74755

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>