[ADD] l10n_nl_reports_vat_pay_wizard: Automatically compute VAT payment reference for Dutch users
The VAT payment reference in the Netherlands changes with each return
and can be computed using a specific formula.

This commit enables Odoo to automatically generate and propose the correct reference
by default when users pay their VAT settlement from the dashboard.

This feature streamlines the process for accounting firms,
reducing the need to repeatedly log in to the Dutch Tax Authorities portal.

task-4063832

closes odoo/enterprise#74770

Signed-off-by: William André (wan) <wan@odoo.com>