[FIX] pos_preparation_display: send notification on qty change
Currently, when reducing the quantity of a line, the notification for
the cancelled items is not sent

Steps to reproduce:
-------------------
* Open both restaurant and preparation display
* In the restaurant place an order for an item with a quantity of 3
* Send the order to the kitchen
* Change the quantity to 2
* Send the order to the kitchen
> Observation: The notification is not sent to the preparation display

Why the fix:
------------
When reducing the quantity, data looks like this
`{'change': True, 'sound': False, 'category_ids': {}}`. Since `category_ids`
is empty, `preparation_displays` will alsos be empty which explains why
the preparation display does not receive the notification, it is simply not
sent.
https://github.com/odoo/enterprise/blob/42b7b322d6c720de45661e3370397adaf74fe1f4/pos_preparation_display/models/preparation_display_order.py#L29-L40

When augmenting the quantity we add the category here:
https://github.com/odoo/enterprise/blob/681c4bca67e9e312f66d797e45fb7e1c3a10b983/pos_preparation_display/models/pos_order.py#L135

We use the same logig when reducing it. This is already done in more recent
version:
https://github.com/odoo/enterprise/blob/42b7b322d6c720de45661e3370397adaf74fe1f4/pos_preparation_display/models/pos_order.py#L166

opw-4332578

closes odoo/enterprise#74803

Signed-off-by: David Monnom (moda) <moda@odoo.com>