[FIX] pos_preparation_display: prevent order duplication
Before this commit, if the user set a database with two preparation
displays, both displaying the same products and orders, duplication
of the order could happen.

Steps to reproduce the issue:
1. configure a pos_restaurant
2. configure 2 kitchen displays (disp1 and disp2)
3. Open the restaurant and create and order with 3-4 items
(items would belong to the same category, ideally)
4. on disp1, proceed to tick/mark as done 2 items
5. on disp1, press on the header of the order -> this will split
the order, sending the done items to the next stage.
6. back in the POS, create a new order (any table)
7. on disp2, you will see a wrong duplication of the order the was
split at the previous step number 5

The duplication would happen even if the order (and related split)
are marked as done on both disp1 and disp2.

What that means in the client flow is that you could have completely processed
the order at both stations, and have it reappear at one of them.
In practice, since inter-station communication is not always possible,
it could lead to the staff preparing the same order multiple times.
This in turn would incur losses because of the wasted orders and increased delays
in processing other legitimate orders.

After this commit, the other preparation displays (disp2 in my example above)
get notified immediately of the change and update their respective statuses to
reflect the order split.

opw-4367937

closes odoo/enterprise#74816

Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>