[FIX] planning: fix traceback when the updating planning slot with no recurrency
Currently, a traceback is occurring when the user tries to update the
planning slot having no recurrence.

To reproduce this issue:

1) Install planning
2) Open any existing Open Shift planning slot
3) Enable the `repeat` and update the start and end date
4) Give the edit value as `All shifts` and make sure the resource be empty
5) Save the record.

Error:- "IndexError: tuple index out of range"

Here the resource is not required in planning slot. So when the
user tries to update a open shift having repeat enabled and
resource_update as `all`, it leads to a traceback from the below line.

https://github.com/odoo/enterprise/blob/64fc38a80520cfeeb81c0bb329c6b24c3e9454e1/planning/models/planning.py#L866-L871

This is because when there is no `recurrency_id` in the slot, we get an empty
recordset for the `planning_slot` and which leads to the above traceback
when try to extract a value from planning_slot.

We can resolve this issue by adding an extra check of recurrency_slots
which makes the code more robust.

sentry-6096445646

closes odoo/enterprise#75134

X-original-commit: 6f3e8e31158dea7b02181cac8c4440b2efe24425
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Altaf Shaik (alsh) <alsh@odoo.com>