[FIX] web_gantt: fix gantt connector accessibility issues
This commit fixes two accessibility issues related to the gantt
view connectors. Prior to this commit, the connectors were using the raw
value when hovering the pill. This is mostly fine is light mode, but for
some specific color, this was sometimes not enough to ensure a good
visual separation between the connectors and the pill. To handle this
issue, we add a pseudo element underneath the border, to make it stand
out and prevent all the potential contrast issues.

There was also a `z-index` issues with the connectors, the one at the
bottom being hidden by the pill below. To fix the issue, we handle the
`z-index` at the pill wrapper level to ensure it gets the right elevation.

task-3865935

closes odoo/enterprise#74859

Signed-off-by: Pierre Paridans (app) <app@odoo.com>