[IMP] account_reports: tax closing send/pay activity
First we replaced the 'send' activity by the 'pay' activity.
Since the 'pay' activity was inadequate for the FIDU /my/databases view (it's only created when there is tax to pay), we brought the 'send' activity back in f39f9c58129c.
But in order to see the 'pay' activity, the 'send' activity had to be marked as done & nobody knows how to do that.
And what if the user wants to pay before sending? (to avoid fines)

So in this PR, we create the 'send' and 'pay' acitivities at the same time.
Those who want to send, send. And those who want to pay, pay.

Task - TSB/BIB/CHKL/AYH discussion

closes odoo/enterprise#75317

X-original-commit: ae63d0c11d425e7f33c6f9034e02c4bab99aa919
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>