[FIX] account_consolidation: handle empty totals when formatting account lines
When processing account lines, the totals list might be empty resulting in an IndexError when trying to set auditable to false on the last column, which leads to consolidated balance to crash.

This fix just makes sure that the modifying the column on this line is only done when the cols actually contains some data.

opw-4205722

closes odoo/enterprise#74938

Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>