[IMP] account_reports: currency table: add test case
If a period containing some res.currency.rate object starts by some sub-period without any range, the average rate computation considered this sub-period as having a rate of 0. It's wrong; it should be 1.

The fix itself is in the corresponding community commit ; this one adds a check ensuring this doesn't break again in the future.

Fixing this also revealed an issue in the Trial Balance, where the initial balance column group was computed with different rates than the rest of the report, which is not what we want for this instance. We fix that by properly setting the period_key on all of its column groups and ensuring the initial balance dates have no impact on the currency table.

closes odoo/enterprise#75140

Related: odoo/odoo#189599
Signed-off-by: John Laterre (jol) <jol@odoo.com>