[FIX] planning: click on dropdown to click on publish button in mobile
Before this commit, when the onboarding tour of planning is launched, a
step to click on the dropdown to get additional buttons next to `New`
button is useless in desktop because the next step is to click on
`Publish` button but that button is displayed next to `New` and the
`Arrow down` button to get the other actions.

This commit adds a condition on the step to click on arrow down button
to get the extra actions in the gantt view of `planning.slot` to only do
that step when the user is in mobile view.

closes odoo/enterprise#74972

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>