[FIX] account: make accounting fields invisble without CoA
When having Accounting installed with a company without chart template,
we can not create a partner as bot `property_account_payable_id` and
`property_account_receivable_id` are required.

With ff54822738925d5b5b952520bbfbd4a76c876809 we add a non stored computed field to handle
the `required` attribute on partner view.
We use this field to make the `Accounting Entries` group invisble when
it is not needed.

opw-4323694

closes odoo/enterprise#75153

X-original-commit: 37cf51aca5cc6bb7dc01d252f28548a1dbe6a411
Related: odoo/odoo#189624
Signed-off-by: John Laterre (jol) <jol@odoo.com>