[FIX] documents: remove tags when moving document
When moving a document from a folder to another, the tags are
not removed, this leads to wrong behavior, for example when grouping
by tags in the destination folder because we have invisible records
linked to the documents.

With this commit, we check for invalid tags and remove them.

Steps:

- With demo data, select the document mail.png in the 'Internal' folder
- Change its folder to 'Finance' via the select input on the right side window
- Go to 'Finance' folder, and filter the documents by tag (tag_ids
  contains 'Inbox')
-> The document mail.png appears in the list although there is no tag
   set on it

opw-4347910

closes odoo/enterprise#78304

X-original-commit: 904c7c8aa99ae64eee3231567799282bce112b8c
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>
Signed-off-by: Guillaume Vanleynseele (guva) <guva@odoo.com>