[FIX] web_grid: better handle step in ranges inside the grid view
Before this commit, if the user alters the grid view to have this kind
of ranges:

```xml
<range name="month" string="Month" span="month" step="month"/>
```
the `step="month"` will not be used to load the data for the grid view
and will continue to load data per day (when this range is the current
one in the grid view).

This commit makes sure the step is correctly used to fetch the data for
each columns displayed in the grid view (in this case only one column
for the whole month should be displayed).

closes odoo/enterprise#75056

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>