[FIX] web_studio: prevent traceback while opening studio
Before this commit:
When the user clicks on the studio icon just after clicking on any app and before its
view gets loaded, an error pops up because at that point there is no valid controller or action
to open studio on.

After this commit there is no error as we prevent entering in studio during the loading of an action/view.

Task-4348166

closes odoo/enterprise#75881

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>