[FIX] pos_order_tracking_display: filter stages from current display
This is a post fix for this commit:
https://github.com/odoo/enterprise/commit/4c6c902651ad48b1064f1eb4eb326fa773b6b098

We need to filter the stages from the current display when we
determine if orders are completed or not since if you change the
stage of an order from the display 1, you will first go through
`_get_pos_orders` with the display 1 and the last stage in
`order_stage_id` is related to display 1. But then you go
a second time in `_get_pos_order` but with display 2 and it
will always says that the order is not ready since the last
stage is related to display 1 and will never be equal to
`second_last_stage`.

opw-4265987

closes odoo/enterprise#75102

Signed-off-by: David Monnom (moda) <moda@odoo.com>