[FIX] account_iso20022: make iso20022_uetr field position consistent
The position of the `iso20022_uetr` field was defined as being after the
`partner_bank_id` field in the view. However multiple of these fields
were defined in the view using if/else conditions, and the former field
was put after the first occurrence of the latter. So depending on which
`partner_bank_id` field was shown, the `iso20022_uetr` field would be
either before or after it.

This commit makes it consistent and puts it after the last occurrence of
`partner_bank_id`.

closes odoo/enterprise#75064

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>