[FIX] web_studio: allow to export custom model
Since [1], the export wizard default export data domain has changed but
it was not at all in the scope of the faulty commit.

This leads to a validation error when a new wizard is created if there
is multiple 'ir.model.data' records created in the context of Studio
and pointing to the same records.
This is often the case for instance for 'ir.model.inherit' records.

This commit reintroduces the missing domain part, while adding a test.

[1]: 99f3591316891e1c2373ed0b3b7e786c0a6b8457

closes odoo/enterprise#75078

Signed-off-by: Bruno Boi (boi) <boi@odoo.com>