[IMP] sale_subscription: decrease the probability of concurrent updates
The logic for _handle_post_invoice_hook_exception is done in sale_subscription_stock and not in sale_subscription.
We do this to prevent calling any code in _post_invoice_hook that is not meant to be called manually.

Prevent concurrent update between crons 'payment: post-process transactions' and 'Sale Subscription: generate recurring invoices and payments' by filtering out subscription with 'is_invoice_cron' = True.

Backport of #70878

If the delivery creations do fail:
 - The error is caught as to not prevent the next batch from being triggered.
 - An exception activity is created on the subscription to notify the customer about the error.
 - A contextual action is available to manually trigger the delivery.

OPW-4319019

closes odoo/enterprise#75618

X-original-commit: d50d953a8c7fe38a26858871a66807aa63b120ad
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: David Fesquet (dafr) <dafr@odoo.com>